requirements.txt
fastapi==0.85.2
geopy==2.2.0
nltk==3.7
numpy==1.23.4
pydantic==1.9.2
redis==4.3.4
redis_om==0.1.0
